+++ packages/oc-docs/src/components/Topbar/Brand/Brand.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import InitialsAvatar from '../../../components/InitialsAvatar/InitialsAvatar';
+import InitialsAvatar from '../../InitialsAvatar/InitialsAvatar';
 import { StyledWrapper } from './StyledWrapper';
 
 export interface BrandProps {
