+++ packages/oc-docs/src/components/Topbar/Topbar.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { StyledWrapper } from './StyledWrapper';
 import Brand from './Brand/Brand';
 import MobileOverflow from './MobileOverflow/MobileOverflow';
-import OpenInBrunoButton from '../../components/OpenInBrunoButton/OpenInBrunoButton';
-import IconButton from '../IconButton/IconButton';
+import OpenInBrunoButton from '../OpenInBrunoButton/OpenInBrunoButton';
+import IconButton from '../../ui/IconButton/IconButton';
 import { SearchIcon, HamburgerIcon } from '../../assets/icons';
 import { useTopbarLayout, showsHamburger } from '../../hooks/useTopbarLayout';
 import { useCanRunBrunoApp } from '../../hooks/useCanRunBrunoApp';
